Move `bigint` module to separate file
